Refactor the whole distributed architecture.
There is now only one scheduler type, standalone, with a simple master/worker model.
For now, the agent encompasses both master/worker capabilities.
Some "core" packages have once again been moved in a "framework" subpackage.
Task repository is not needed anymore.

GitOrigin-RevId: adf4509c71db123ed1d0e9492222f462dea9404c

+++ src/jvm/fr/cnrs/liris/accio/agent/AgentServer.scala
@@ -27,10 +27,11 @@ import com.twitter.finatra.thrift.ThriftServer
 import com.twitter.finatra.thrift.filters._
 import com.twitter.finatra.thrift.routing.ThriftRouter
 import com.twitter.util.FuturePool
+import fr.cnrs.liris.accio.agent.master.LostWorkerObserver
+import fr.cnrs.liris.accio.agent.worker.{LostExecutorsObserver, WorkerLifecycle}
 import fr.cnrs.liris.accio.core.dsl.inject.DslModule
 import fr.cnrs.liris.accio.core.filesystem.inject.FileSystemModule
 import fr.cnrs.liris.accio.core.scheduler.inject.SchedulerModule
-import fr.cnrs.liris.accio.core.statemgr.inject.StateManagerModule
 import fr.cnrs.liris.accio.core.storage.inject.StorageModule
 import fr.cnrs.liris.privamov.ops.OpsModule
 import org.slf4j.LoggerFactory
@@ -42,7 +43,6 @@ class AgentServer extends ThriftServer {
 
   override protected def modules: Seq[Module] = Seq(
     FileSystemModule,
-    StateManagerModule,
     SchedulerModule,
     StorageModule,
     DslModule,
@@ -60,10 +60,22 @@ class AgentServer extends ThriftServer {
   }
 
   override protected def start(): Unit = {
-    val observer = injector.instance[LostTaskObserver]
-    FuturePool.unboundedPool(observer.run())
-    onExit {
-      observer.kill()
+    if (AgentModule.masterFlag()) {
+      val observer = injector.instance[LostWorkerObserver]
+      FuturePool.unboundedPool(observer.run())
+      onExit {
+        observer.kill()
+      }
+    }
+    if (AgentModule.workerFlag()) {
+      val lifecycle = injector.instance[WorkerLifecycle]
+      lifecycle.register()
+      val observer  = injector.instance[LostExecutorsObserver]
+      FuturePool.unboundedPool(observer.run())
+      onExit {
+        lifecycle.unregister()
+        observer.kill()
+      }
     }
   }
